accomplishment

+++ app/Http/Controllers/TasksController.php
@@ -6,7 +6,7 @@ use Illuminate\Http\Request;
 
 use App\Http\Requests;
 
-use App\Task; 
+use App\Task;
 
 class TasksController extends Controller
 {
@@ -17,22 +17,24 @@ class TasksController extends Controller
      */
     public function index()
     {
+     $data = [];
         if (\Auth::check()) {
-        $tasks = Task::all();
-
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            //$tasks = Task::all();
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            
+            return view('tasks.index', $data);
         }else {
             return view('welcome');
-        }
     }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    
+    }
+    
     public function create()
     {
         $task = new Task;
@@ -41,7 +43,6 @@ class TasksController extends Controller
             'task' => $task,
         ]);
     }
-
     /**
      * Store a newly created resource in storage.
      *
@@ -51,12 +52,14 @@ class TasksController extends Controller
     public function store(Request $request)
     {
         $this->validate($request, [
-            'status' => 'required|max:10', 
-            'content' => 'required|max:10',
+            'status' => 'required|max:10',
+            'content' => 'required|max:191',
         ]);
-        $task= new Task;
-        $task->status = $request->status; 
+
+        $task = new Task;
+        $task->status = $request->status;  
         $task->content = $request->content;
+        $task->user_id = \Auth::user()->id;
         $task->save();
 
         return redirect('/');
@@ -76,7 +79,6 @@ class TasksController extends Controller
             'task' => $task,
         ]);
     }
-
     /**
      * Show the form for editing the specified resource.
      *
@@ -85,7 +87,7 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task = Task::find($id);
+    $task = Task::find($id);
 
         return view('tasks.edit', [
             'task' => $task,
@@ -101,17 +103,19 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
-         $this->validate($request, [
-             'status' => 'required|max:10',
-            'content' => 'required|max:10',
+        $this->validate($request, [
+            'status' => 'required|max:10', 
+            'content' => 'required|max:191',
         ]);
+
         
         $task = Task::find($id);
-        $task->status = $request->status; 
+        $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
 
         return redirect('/');
+
     }
 
     /**
@@ -122,8 +126,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-         $task = Task::find($id);
-        $task->delete();
+        $task = Task::find($id);
+        
+        
+         if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
 
         return redirect('/');
     }
